BUGFIX: QA-5047 - Before manipulating with GitHub repository test whether the object is initialised
The object for manipulation with the GitHub repository is intialised
by default when the Trigger is created or the Jenkins is started. But
when in time of trigger creation the Jenkins does not have permission
to pull the repository, the object is not created and the job does
not work properly and fails every time when accessing the repo even
after the permission has been fixed.